Reorganize FreeBSD code.
